+++ src/Illuminate/Contracts/Queue/Resumable.php
@@ -2,8 +2,8 @@
 
 namespace Illuminate\Contracts\Queue;
 
-use Illuminate\Bus\Workflow\ExecutionState;
-use Illuminate\Bus\Workflow\JobSequence;
+use Illuminate\Bus\JobSequence\ExecutionState;
+use Illuminate\Bus\JobSequence\JobSequence;
 
 interface Resumable
 {
